backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ─── Authentication ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ─── Validation & Settings ────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.27.0

# ─── Caching & Rate Limiting ──────────────────────────────────────────────────
redis==5.0.4
slowapi==0.1.9

# ─── Storage ──────────────────────────────────────────────────────────────────
boto3==1.34.99

# ─── Payments ─────────────────────────────────────────────────────────────────
razorpay==1.4.1
setuptools==69.5.1

# ─── Async & Background Tasks ─────────────────────────────────────────────────
celery==5.4.0
flower==2.0.1

# ─── Logging ──────────────────────────────────────────────────────────────────
structlog==24.1.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
factory-boy==3.3.0

# ─── Dev / Linting ────────────────────────────────────────────────────────────
ruff==0.4.4
mypy==1.10.0
pre-commit==3.7.1
